Move DashboardNavigation components into a folder.
